+++ day5/演習3/tests/test_model.py
@@ -1,4 +1,7 @@
+import gc
 import os
+
+import psutil
 import pytest
 import pandas as pd
 import numpy as np
@@ -171,3 +174,21 @@ def test_model_reproducibility(sample_data, preprocessor):
     assert np.array_equal(
         predictions1, predictions2
     ), "モデルの予測結果に再現性がありません"
+
+def test_model_memory_usage(train_model):
+    """モデルのメモリ使用量を検証"""
+    model, X_test, _ = train_model
+    
+    process = psutil.Process(os.getpid())
+    gc.collect()
+    
+    memory_before = process.memory_info().rss / (1024 * 1024)
+    model.predict(X_test)
+    memory_after = process.memory_info().rss / (1024 * 1024)
+    
+    memory_used = memory_after - memory_before
+    
+    memory_threshold = 100
+    
+    print(f"メモリ使用量: {memory_used:.2f} MB")
+    assert memory_used < memory_threshold, f"メモリ使用量が多すぎます: {memory_used:.2f} MB (閾値: {memory_threshold} MB)"
